Cache the nix store database too
Nix records all store paths in its database. If Nix evaluates an
expression which refers to a store path not in the Nix database, it will
rebuilt the corresponding derivation. Hence, such paths aren't really
cached. The garbage collector also checks the validity of store paths
(and requires the database to find dependencies) and deletes all invalid
paths.