Reduce the size of the Nix store before caching it
When caching a derivation which requires a lot of build dependencies but
doesn't require them at runtime, they are still included in GitHub's
cache. Hence, the garbage collector is run to reduce the cache size and
thus enhance the cache restore performance. For the same reason, the
cache is optimized by using hard links where appropriate (`nix-store`s
`--optimise` flag).