Cache the Nix database too (#20)
* Cache the nix store database too

Nix records all store paths in its database. If Nix evaluates an
expression which refers to a store path not in the Nix database, it will
rebuilt the corresponding derivation. Hence, such paths aren't really
cached. The garbage collector also checks the validity of store paths
(and requires the database to find dependencies) and deletes all invalid
paths.

* Reduce the size of the Nix store before caching it

When caching a derivation which requires a lot of build dependencies but
doesn't require them at runtime, they are still included in GitHub's
cache. Hence, the garbage collector is run to reduce the cache size and
thus enhance the cache restore performance. For the same reason, the
cache is optimized by using hard links where appropriate (`nix-store`s
`--optimise` flag).

* Bump the GitHub action dependency versions

* Bump Nix version